+++ sdk/keyvault/azure-keyvault-certificates/tests/test_certificates_client_async.py
@@ -602,13 +602,15 @@ class CertificateClientTests(CertificatesTestCase, KeyVaultTestCase):
 
         for message in mock_handler.messages:
             if message.levelname == "DEBUG" and message.funcName == "on_request":
-                try:
-                    body = json.loads(message.message)
-                    if body["provider"] == "Test":
-                        return
-                except (ValueError, KeyError):
-                    # this means the message is not JSON or has no kty property
-                    pass
+                messages_request = message.message.split("/n")
+                for m in messages_request:
+                    try:
+                        body = json.loads(m)
+                        if body["provider"] == "Test":
+                            return
+                    except (ValueError, KeyError):
+                        # this means the message is not JSON or has no kty property
+                        pass
 
         assert False, "Expected request body wasn't logged"
 
@@ -626,12 +628,14 @@ class CertificateClientTests(CertificatesTestCase, KeyVaultTestCase):
 
         for message in mock_handler.messages:
             if message.levelname == "DEBUG" and message.funcName == "on_request":
-                try:
-                    body = json.loads(message.message)
-                    assert body["provider"] != "Test", "Client request body was logged"
-                except (ValueError, KeyError):
-                    # this means the message is not JSON or has no kty property
-                    pass
+                messages_request = message.message.split("/n")
+                for m in messages_request:
+                    try:
+                        body = json.loads(m)
+                        assert body["provider"] != "Test", "Client request body was logged"
+                    except (ValueError, KeyError):
+                        # this means the message is not JSON or has no kty property
+                        pass
 
     @all_api_versions()
     @client_setup