+++ sdk/monitor/azure-monitor-query/azure/monitor/query/_logs_query_client.py
@@ -5,16 +5,22 @@
 # license information.
 # --------------------------------------------------------------------------
 
-from typing import TYPE_CHECKING, Any, Union, Sequence, Dict, List
+from typing import TYPE_CHECKING, Any, Union, Sequence, Dict, List, cast
 from azure.core.exceptions import HttpResponseError
 from azure.core.tracing.decorator import distributed_trace
 
 from ._generated._monitor_query_client import MonitorQueryClient
 
 from ._generated.models import BatchRequest, QueryBody as LogsQueryBody
-from ._helpers import get_authentication_policy, construct_iso8601, order_results, process_error, process_prefer
-from ._models import LogsBatchQuery, LogsQueryResult
-from ._exceptions import  LogsQueryError, QueryPartialErrorException
+from ._helpers import (
+    get_authentication_policy,
+    construct_iso8601,
+    order_results,
+    process_error,
+    process_prefer,
+)
+from ._models import LogsBatchQuery, LogsQueryResult, LogsQueryPartialResult
+from ._exceptions import LogsQueryError
 
 if TYPE_CHECKING:
     from azure.core.credentials import TokenCredential
@@ -22,11 +28,17 @@ if TYPE_CHECKING:
 
 
 class LogsQueryClient(object):
-    """LogsQueryClient
+    """LogsQueryClient. Use this client to collect and organize log and performance data from
+    monitored resources. Data from different sources such as platform logs from Azure services,
+    log and performance data from virtual machines agents, and usage and performance data from
+    apps can be consolidated into a single Azure Log Analytics workspace.
+
+    The various data types can be analyzed together using the
+    [Kusto Query Language](https://docs.microsoft.com/azure/data-explorer/kusto/query/)
 
     .. admonition:: Example:
 
-    .. literalinclude:: ../samples/sample_log_query_client.py
+    .. literalinclude:: ../samples/sample_single_logs_query.py
         :start-after: [START client_auth_with_token_cred]
         :end-before: [END client_auth_with_token_cred]
         :language: python
@@ -42,7 +54,7 @@ class LogsQueryClient(object):
     def __init__(self, credential, **kwargs):
         # type: (TokenCredential, Any) -> None
 
-        self._endpoint = kwargs.pop('endpoint', 'https://api.loganalytics.io/v1')
+        self._endpoint = kwargs.pop("endpoint", "https://api.loganalytics.io/v1")
         self._client = MonitorQueryClient(
             credential=credential,
             authentication_policy=get_authentication_policy(credential),
@@ -52,8 +64,8 @@ class LogsQueryClient(object):
         self._query_op = self._client.query
 
     @distributed_trace
-    def query(self, workspace_id, query, **kwargs):
-        # type: (str, str, Any) -> LogsQueryResult
+    def query_workspace(self, workspace_id, query, **kwargs):
+        # type: (str, str, Any) -> Union[LogsQueryResult, LogsQueryPartialResult]
         """Execute an Analytics query.
 
         Executes an Analytics query for data.
@@ -64,7 +76,7 @@ class LogsQueryClient(object):
         :param query: The Kusto query. Learn more about the `Kusto query syntax
          <https://docs.microsoft.com/azure/data-explorer/kusto/query/>`_.
         :type query: str
-        :keyword timespan: The timespan for which to query the data. This can be a timedelta,
+        :keyword timespan: Required. The timespan for which to query the data. This can be a timedelta,
          a timedelta and a start datetime, or a start datetime/end datetime.
         :paramtype timespan: ~datetime.timedelta or tuple[~datetime.datetime, ~datetime.timedelta]
          or tuple[~datetime.datetime, ~datetime.datetime]
@@ -77,73 +89,76 @@ class LogsQueryClient(object):
         :keyword additional_workspaces: A list of workspaces that are included in the query.
          These can be qualified workspace names, workspace Ids, or Azure resource Ids.
         :paramtype additional_workspaces: list[str]
-        :keyword allow_partial_errors: Defaults to False. If set to true, partial errors are not thrown.
-        :paramtype allow_partial_errors: bool
-        :return: LogsQueryResult, or the result of cls(response)
-        :rtype: ~azure.monitor.query.LogsQueryResult
+        :return: LogsQueryResult if there is a success or LogsQueryPartialResult when there is a partial success.
+        :rtype: Union[~azure.monitor.query.LogsQueryResult, ~azure.monitor.query.LogsQueryPartialResult]
         :raises: ~azure.core.exceptions.HttpResponseError
 
         .. admonition:: Example:
 
-        .. literalinclude:: ../samples/sample_log_query_client.py
+        .. literalinclude:: ../samples/sample_single_logs_query.py
             :start-after: [START send_logs_query]
             :end-before: [END send_logs_query]
             :language: python
             :dedent: 0
             :caption: Get a response for a single Log Query
         """
-        allow_partial_errors = kwargs.pop('allow_partial_errors', False)
-        if 'timespan' not in kwargs:
-            raise TypeError("query() missing 1 required keyword-only argument: 'timespan'")
-        timespan = construct_iso8601(kwargs.pop('timespan'))
+        if "timespan" not in kwargs:
+            raise TypeError(
+                "query() missing 1 required keyword-only argument: 'timespan'"
+            )
+        timespan = construct_iso8601(kwargs.pop("timespan"))
         include_statistics = kwargs.pop("include_statistics", False)
         include_visualization = kwargs.pop("include_visualization", False)
         server_timeout = kwargs.pop("server_timeout", None)
         workspaces = kwargs.pop("additional_workspaces", None)
 
-        prefer = process_prefer(server_timeout, include_statistics, include_visualization)
+        prefer = process_prefer(
+            server_timeout, include_statistics, include_visualization
+        )
 
         body = LogsQueryBody(
-            query=query,
-            timespan=timespan,
-            workspaces=workspaces,
-            **kwargs
+            query=query, timespan=timespan, workspaces=workspaces, **kwargs
         )
 
         try:
-            generated_response = self._query_op.execute( # pylint: disable=protected-access
-                workspace_id=workspace_id,
-                body=body,
-                prefer=prefer,
-                **kwargs
+            generated_response = (
+                self._query_op.execute(  # pylint: disable=protected-access
+                    workspace_id=workspace_id, body=body, prefer=prefer, **kwargs
+                )
             )
         except HttpResponseError as err:
             process_error(err, LogsQueryError)
-        response = LogsQueryResult._from_generated(generated_response) # pylint: disable=protected-access
+        response = None
         if not generated_response.error:
-            return response
-        if not allow_partial_errors:
-            raise QueryPartialErrorException(error=generated_response.error)
-        response.partial_error = LogsQueryError._from_generated( # pylint: disable=protected-access
-            generated_response.error
+            response = LogsQueryResult._from_generated( # pylint: disable=protected-access
+                generated_response
             )
-        return response
+        else:
+            response = LogsQueryPartialResult._from_generated( # pylint: disable=protected-access
+                generated_response, LogsQueryError
+            )
+        return cast(Union[LogsQueryResult, LogsQueryPartialResult], response)
 
     @distributed_trace
-    def query_batch(self, queries, **kwargs):
-        # type: (Union[Sequence[Dict], Sequence[LogsBatchQuery]], Any) -> List[LogsQueryResult]
-        """Execute a list of analytics queries. Each request can be either a LogQueryRequest
+    def query_batch(
+        self,
+        queries,  # type: Union[Sequence[Dict], Sequence[LogsBatchQuery]]
+        **kwargs  # type: Any
+    ):
+        # type: (...) -> List[Union[LogsQueryResult, LogsQueryPartialResult, LogsQueryError]]
+        """Execute a list of analytics queries. Each request can be either a LogsBatchQuery
         object or an equivalent serialized model.
 
-        The response is returned in the same order as that of the requests sent.
+        **NOTE**: The response is returned in the same order as that of the requests sent.
 
         :param queries: The list of Kusto queries to execute.
         :type queries: list[dict] or list[~azure.monitor.query.LogsBatchQuery]
-        :keyword bool allow_partial_errors: If set to True, a `LogsQueryResult` object is returned
-         when a partial error occurs. The error can be accessed using the `partial_error`
-         attribute in the object.
-        :return: List of LogsQueryResult, or the result of cls(response)
-        :rtype: list[~azure.monitor.query.LogsQueryResult]
+        :return: List of LogsQueryResult, LogsQueryPartialResult and LogsQueryError.
+         For a given query, a LogsQueryResult is returned if the response is a success, LogsQueryPartialResult
+         is returned when there is a partial success and a LogsQueryError is returned when there is a failure.
+         The status of each response can be checked using `LogsQueryStatus` enum.
+        :rtype: list[Union[~azure.monitor.query.LogsQueryResult, ~azure.monitor.query.LogsQueryPartialResult,
+         ~azure.monitor.query.LogsQueryError]
         :raises: ~azure.core.exceptions.HttpResponseError
 
         .. admonition:: Example:
@@ -155,25 +170,28 @@ class LogsQueryClient(object):
             :dedent: 0
             :caption: Get a response for multiple Log Queries.
         """
-        allow_partial_errors = kwargs.pop('allow_partial_errors', False)
         try:
-            queries = [LogsBatchQuery(**q) for q in queries]
+            queries = [LogsBatchQuery(**cast(Dict, q)) for q in queries]
         except (KeyError, TypeError):
             pass
-        queries = [q._to_generated() for q in queries] # pylint: disable=protected-access
+        queries = [
+            cast(LogsBatchQuery, q)._to_generated() for q in queries # pylint: disable=protected-access
+        ]
         try:
             request_order = [req.id for req in queries]
         except AttributeError:
-            request_order = [req['id'] for req in queries]
+            request_order = [req["id"] for req in queries]
         batch = BatchRequest(requests=queries)
         generated = self._query_op.batch(batch, **kwargs)
-        mapping = {item.id: item for item in generated.responses}
+        mapping = {item.id: item for item in generated.responses} # type: ignore
         return order_results(
             request_order,
             mapping,
-            LogsQueryResult,
-            LogsQueryError,
-            allow_partial_errors)
+            obj=LogsQueryResult,
+            err=LogsQueryError,
+            partial_err=LogsQueryPartialResult,
+            raise_with=LogsQueryError,
+        )
 
     def close(self):
         # type: () -> None