+++ sdk/monitor/azure-monitor-query/tests/async/test_exceptions_async.py
@@ -3,7 +3,7 @@ import pytest
 import os
 from azure.identity.aio import ClientSecretCredential
 from azure.core.exceptions import HttpResponseError
-from azure.monitor.query import LogsBatchQuery, LogsQueryError,LogsQueryResult, QueryPartialErrorException
+from azure.monitor.query import LogsBatchQuery, LogsQueryError,LogsQueryResult, LogsQueryPartialResult
 from azure.monitor.query.aio import LogsQueryClient
 
 def _credential():
@@ -20,7 +20,7 @@ async def test_logs_single_query_fatal_exception():
     credential = _credential()
     client = LogsQueryClient(credential)
     with pytest.raises(HttpResponseError):
-        await client.query('bad_workspace_id', 'AppRequests', timespan=None)
+        await client.query_workspace('bad_workspace_id', 'AppRequests', timespan=None)
 
 @pytest.mark.live_test_only
 @pytest.mark.asyncio
@@ -30,18 +30,8 @@ async def test_logs_single_query_partial_exception_not_allowed():
     query = """let Weight = 92233720368547758;
     range x from 1 to 3 step 1
     | summarize percentilesw(x, Weight * 100, 50)"""
-    with pytest.raises(QueryPartialErrorException) as err:
-        await client.query(os.environ['LOG_WORKSPACE_ID'], query, timespan=timedelta(days=1))
-
-@pytest.mark.live_test_only
-@pytest.mark.asyncio
-async def test_logs_single_query_partial_exception_allowed():
-    credential = _credential()
-    client = LogsQueryClient(credential)
-    query = """let Weight = 92233720368547758;
-    range x from 1 to 3 step 1
-    | summarize percentilesw(x, Weight * 100, 50)"""
-    response = await client.query(os.environ['LOG_WORKSPACE_ID'], query, timespan=timedelta(days=1), allow_partial_errors=True)
+    response = await client.query_workspace(os.environ['LOG_WORKSPACE_ID'], query, timespan=timedelta(days=1))
+    assert response.__class__ == LogsQueryPartialResult
     assert response.partial_error is not None
     assert response.partial_error.code == 'PartialError'
     assert response.partial_error.__class__ == LogsQueryError
@@ -76,7 +66,7 @@ async def test_logs_batch_query_fatal_exception():
         ),
     ]
     with pytest.raises(HttpResponseError):
-        await client.query_batch(requests, allow_partial_errors=True)
+        await client.query_batch(requests)
 
 @pytest.mark.live_test_only
 @pytest.mark.asyncio
@@ -90,7 +80,7 @@ async def test_logs_batch_query_partial_exception_not_allowed():
             workspace_id= os.environ['LOG_WORKSPACE_ID']
         ),
         LogsBatchQuery(
-            query= """AppRequests | take 10""",
+            query= """bad query | take 10""",
             timespan=(datetime(2021, 6, 2), timedelta(days=1)),
             workspace_id= os.environ['LOG_WORKSPACE_ID']
         ),
@@ -106,66 +96,5 @@ async def test_logs_batch_query_partial_exception_not_allowed():
     responses = await client.query_batch(requests)
     r1, r2, r3 = responses[0], responses[1], responses[2]
     assert r1.__class__ == LogsQueryResult
-    assert r2.__class__ == LogsQueryResult
-    assert r3.__class__ == LogsQueryError
-
-@pytest.mark.live_test_only
-@pytest.mark.asyncio
-async def test_logs_batch_query_partial_exception_allowed():
-    credential = _credential()
-    client = LogsQueryClient(credential)
-    requests = [
-        LogsBatchQuery(
-            query="AzureActivity | summarize count()",
-            timespan=timedelta(hours=1),
-            workspace_id= os.environ['LOG_WORKSPACE_ID']
-        ),
-        LogsBatchQuery(
-            query= """AppRequests | take 10""",
-            timespan=(datetime(2021, 6, 2), timedelta(days=1)),
-            workspace_id= os.environ['LOG_WORKSPACE_ID']
-        ),
-        LogsBatchQuery(
-            query= """let Weight = 92233720368547758;
-            range x from 1 to 3 step 1
-            | summarize percentilesw(x, Weight * 100, 50)""",
-            workspace_id= os.environ['LOG_WORKSPACE_ID'],
-            timespan=(datetime(2021, 6, 2), datetime(2021, 6, 3)),
-            include_statistics=True
-        ),
-    ]
-    responses = await client.query_batch(requests, allow_partial_errors=True)
-    r1, r2, r3 = responses[0], responses[1], responses[2]
-    assert r1.__class__ == LogsQueryResult
-    assert r2.__class__ == LogsQueryResult
-    assert r3.__class__ == LogsQueryResult
-    assert r3.partial_error is not None
-
-@pytest.mark.live_test_only
-@pytest.mark.asyncio
-async def test_logs_batch_query_non_fatal_exception():
-    credential = _credential()
-    client = LogsQueryClient(credential)
-    requests = [
-        LogsBatchQuery(
-            query="AzureActivity | summarize count()",
-            timespan=timedelta(hours=1),
-            workspace_id= os.environ['LOG_WORKSPACE_ID']
-        ),
-        LogsBatchQuery(
-            query= """AppRequests | take 10""",
-            timespan=(datetime(2021, 6, 2), timedelta(days=1)),
-            workspace_id= os.environ['LOG_WORKSPACE_ID']
-        ),
-        LogsBatchQuery(
-            query= """Bad Query""",
-            workspace_id= os.environ['LOG_WORKSPACE_ID'],
-            timespan=(datetime(2021, 6, 2), datetime(2021, 6, 3)),
-            include_statistics=True
-        ),
-    ]
-    responses = await client.query_batch(requests)
-    r1, r2, r3 = responses[0], responses[1], responses[2]
-    assert r1.__class__ == LogsQueryResult
-    assert r2.__class__ == LogsQueryResult
-    assert r3.__class__ == LogsQueryError
+    assert r2.__class__ == LogsQueryError
+    assert r3.__class__ == LogsQueryPartialResult