+++ sdk/schemaregistry/azure-schemaregistry/azure/schemaregistry/_schema_registry_client.py
@@ -29,9 +29,11 @@ from ._common._constants import SerializationType
 from ._common._schema import Schema, SchemaProperties
 from ._common._response_handlers import (
     _parse_response_schema,
-    _parse_response_schema_id,
+    _parse_response_schema_properties,
 )
 from ._generated._azure_schema_registry import AzureSchemaRegistry
+from ._generated.rest import schema as schema_rest
+
 
 if TYPE_CHECKING:
     from azure.core.credentials import TokenCredential
@@ -96,6 +98,8 @@ class SchemaRegistryClient(object):
          For now Avro is the only supported serialization type by the service.
         :type serialization_type: Union[str, SerializationType]
         :param str schema_content: String representation of the schema being registered.
+        :keyword content_type: The content type of the request. Default value is 'application/json'.
+        :paramtype content_type: str
         :rtype: SchemaProperties
 
         .. admonition:: Example:
@@ -113,14 +117,19 @@ class SchemaRegistryClient(object):
         except AttributeError:
             pass
 
-        schema_properties = self._generated_client.schema.register(
+        request = schema_rest.build_register_request(
             group_name=schema_group,
             schema_name=schema_name,
-            schema_content=schema_content,
-            x_schema_type=serialization_type,
-            cls=_parse_response_schema_id,
+            content=schema_content,
+            serialization_type=serialization_type,
+            content_type=kwargs.pop("content_type", "application/json"),
             **kwargs
         )
+
+        response = self._generated_client.send_request(request)
+        response.raise_for_status()
+        schema_properties = _parse_response_schema_properties(response)
+
         schema_description = (
             schema_group,
             schema_name,
@@ -156,9 +165,10 @@ class SchemaRegistryClient(object):
         try:
             return self._id_to_schema[schema_id]
         except KeyError:
-            schema = self._generated_client.schema.get_by_id(
-                schema_id, cls=_parse_response_schema, **kwargs
-            )
+            request = schema_rest.build_get_by_id_request(schema_id=schema_id)
+            response = self._generated_client.send_request(request, **kwargs)
+            response.raise_for_status()
+            schema = _parse_response_schema(response)
             self._id_to_schema[schema_id] = schema
             return schema
 
@@ -173,9 +183,10 @@ class SchemaRegistryClient(object):
         :param str schema_group: Schema group under which schema should be registered.
         :param str schema_name: Name of schema being registered.
         :param serialization_type: Serialization type for the schema being registered.
-         The
         :type serialization_type: Union[str, SerializationType]
         :param str schema_content: String representation of the schema being registered.
+        :keyword content_type: The content type of the request. Default value is 'application/json'.
+        :paramtype content_type: str
         :rtype: SchemaProperties
 
         .. admonition:: Example:
@@ -199,14 +210,19 @@ class SchemaRegistryClient(object):
             ]
             return properties
         except KeyError:
-            schema_properties = self._generated_client.schema.query_id_by_content(
+            request = schema_rest.build_query_id_by_content_request(
                 group_name=schema_group,
                 schema_name=schema_name,
-                schema_content=schema_content,
-                x_schema_type=serialization_type,
-                cls=_parse_response_schema_id,
+                content=schema_content,
+                serialization_type=serialization_type,
+                content_type=kwargs.pop("content_type", "application/json"),
                 **kwargs
             )
+
+            response = self._generated_client.send_request(request, **kwargs)
+            response.raise_for_status()
+            schema_properties = _parse_response_schema_properties(response)
+
             if not self._id_to_schema.get(schema_properties.schema_id):
                 self._id_to_schema[schema_properties.schema_id] = Schema(schema_content, schema_properties)
             else: