+++ sdk/schemaregistry/azure-schemaregistry-avroserializer/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Release History
 
+## 1.0.0b3 (Unreleased)
+
+### Features Added
+
+### Breaking Changes
+
+### Bugs Fixed
+
+### Other Changes
+
 ## 1.0.0b2 (2021-08-18)
 
 This version and all future versions will require Python 2.7 or Python 3.6+, Python 3.5 is no longer supported.